Organized statistics directory structure
