+++ src/main/java/seedu/billboard/logic/Logic.java
@@ -11,7 +11,7 @@ import seedu.billboard.logic.parser.exceptions.ParseException;
 import seedu.billboard.model.ReadOnlyArchiveWrapper;
 import seedu.billboard.model.ReadOnlyBillboard;
 import seedu.billboard.model.expense.Expense;
-import seedu.billboard.model.statistics.StatisticsType;
+import seedu.billboard.model.statistics.formats.StatisticsFormat;
 
 /**
  * API of the Logic component
@@ -39,7 +39,7 @@ public interface Logic {
     /**
      * Returns an observable view of the selected statistic type.
      */
-    ObservableData<StatisticsType> getStatisticsType();
+    ObservableData<StatisticsFormat> getStatisticsType();
 
     /**
      * Returns the user prefs' billboard file path.
