+++ src/main/java/seedu/billboard/logic/LogicManager.java
@@ -17,7 +17,7 @@ import seedu.billboard.model.Model;
 import seedu.billboard.model.ReadOnlyArchiveWrapper;
 import seedu.billboard.model.ReadOnlyBillboard;
 import seedu.billboard.model.expense.Expense;
-import seedu.billboard.model.statistics.StatisticsType;
+import seedu.billboard.model.statistics.formats.StatisticsFormat;
 import seedu.billboard.storage.Storage;
 
 /**
@@ -65,7 +65,7 @@ public class LogicManager implements Logic {
     }
 
     @Override
-    public ObservableData<StatisticsType> getStatisticsType() {
+    public ObservableData<StatisticsFormat> getStatisticsType() {
         return model.getStatisticsType();
     }
