--- src/main/java/seedu/billboard/model/statistics/StatisticsType.java
@@ -1,8 +1,0 @@
-package seedu.billboard.model.statistics;
-
-/**
- * Enum representing the supported statistics types.
- */
-public enum StatisticsType {
-    TIMELINE
-}
