+++ src/main/java/seedu/billboard/model/statistics/formats/StatisticsFormat.java
@@ -1,0 +1,9 @@
+package seedu.billboard.model.statistics.formats;
+
+/**
+ * Enum representing the supported statistics types.
+ */
+public enum StatisticsFormat {
+    TIMELINE,
+    BREAKDOWN
+}
