+++ src/main/java/seedu/billboard/ui/charts/ChartBox.java
@@ -7,8 +7,8 @@ import javafx.scene.layout.Region;
 import seedu.billboard.commons.core.date.DateInterval;
 import seedu.billboard.commons.core.observable.ObservableData;
 import seedu.billboard.model.expense.Expense;
-import seedu.billboard.model.statistics.TimelineGenerator;
-import seedu.billboard.model.statistics.StatisticsType;
+import seedu.billboard.model.statistics.generators.TimelineGenerator;
+import seedu.billboard.model.statistics.formats.StatisticsFormat;
 import seedu.billboard.ui.UiPart;
 
 
@@ -26,7 +26,7 @@ public class ChartBox extends UiPart<Region> {
     private ObservableData<DateInterval> dateInterval;
     private ExpenseChart currentChart;
 
-    public ChartBox(ObservableData<StatisticsType> statsType, ObservableList<Expense> expenses) {
+    public ChartBox(ObservableData<StatisticsFormat> statsType, ObservableList<Expense> expenses) {
         super(FXML);
 
         this.expenses = expenses;
@@ -40,7 +40,7 @@ public class ChartBox extends UiPart<Region> {
      * Callback which is called when the observed statistics type changes.
      * @param type New statistic type to display.
      */
-    private void onStatsTypeChanged(StatisticsType type) {
+    private void onStatsTypeChanged(StatisticsFormat type) {
         if (currentChart != null) {
             chartContainer.getChildren().remove(currentChart.getRoot());
         }
@@ -48,6 +48,9 @@ public class ChartBox extends UiPart<Region> {
         switch (type) {
         case TIMELINE:
             currentChart = new ExpenseTimelineChart(expenses, dateInterval, new TimelineGenerator());
+            break;
+        case BREAKDOWN:
+            currentChart = null; // TODO
             break;
         default:
             throw new UnsupportedOperationException("Chart not implemented for selected statistic");
