+++ src/main/resources/view/ExpenseBreakdownChart.fxml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.layout.*?>
+
+<?import javafx.scene.chart.PieChart?>
+<AnchorPane xmlns="http://javafx.com/javafx"
+            xmlns:fx="http://javafx.com/fxml">
+
+    <PieChart fx:id="breakdownChart"
+              AnchorPane.bottomAnchor="0.0"
+              AnchorPane.leftAnchor="0.0"
+              AnchorPane.rightAnchor="0.0"
+              AnchorPane.topAnchor="0.0"
+              minHeight="350.0" prefHeight="350.0" maxHeight="350.0"
+              minWidth="540.0" prefWidth="540.0" maxWidth="540.0"
+              title="Breakdown"
+              legendSide="LEFT">
+
+    </PieChart>
+
+</AnchorPane>
